+++ src/tk_uia/__init__.py
@@ -39,7 +39,7 @@ from tk_uia.tkversion import Strategy
 if TYPE_CHECKING:
     import tkinter
 
-__version__ = "0.6.0"
+__version__ = "0.6.1"
 
 __all__ = [
     "ROLE_FOR_TK_CLASS",